Startopia — Uppdate script to 2.11 syntax

+++ play.it-2/games/play-startopia.sh
@@ -1,8 +1,8 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
-# Copyright (c) 2015-2019, Antoine "vv221/vv222" Le Gonidec
+# Copyright (c) 2015-2020, Antoine "vv221/vv222" Le Gonidec
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -30,18 +30,16 @@ set -o errexit
 
 ###
 # Startopia
-# build native Linux packages from the original installers
-# send your bug reports to vv221@dotslashplay.it
+# build native packages from the original installers
+# send your bug reports to contact@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20200210.1
 
 # Set game-specific variables
 
 GAME_ID='startopia'
 GAME_NAME='Startopia'
-
-ARCHIVES_LIST='ARCHIVE_GOG'
 
 ARCHIVE_GOG='setup_startopia_2.0.0.17.exe'
 ARCHIVE_GOG_URL='https://www.gog.com/game/startopia'
@@ -49,17 +47,14 @@ ARCHIVE_GOG_MD5='4fe8d194afc1012e136ed3e82f1de171'
 ARCHIVE_GOG_VERSION='1.01b-gog2.0.0.17'
 ARCHIVE_GOG_SIZE='600000'
 
-ARCHIVE_DOC1_PATH='tmp'
-ARCHIVE_DOC1_FILES='./*eula.txt'
-
-ARCHIVE_DOC2_PATH='app'
-ARCHIVE_DOC2_FILES='./eula ./weblinks ./*.html ./*.pdf ./*.rtf ./*.txt'
+ARCHIVE_DOC_DATA_PATH='app'
+ARCHIVE_DOC_DATA_FILES='eula weblinks *.html *.pdf *.rtf *.txt'
 
 ARCHIVE_GAME_BIN_PATH='app'
-ARCHIVE_GAME_BIN_FILES='./binkw32.dll ./startopia.exe ./startopia.ini'
+ARCHIVE_GAME_BIN_FILES='binkw32.dll startopia.exe startopia.ini'
 
 ARCHIVE_GAME_DATA_PATH='app'
-ARCHIVE_GAME_DATA_FILES='./cardid.tom ./data ./intro ./languageinis ./missions ./startopia.jpg ./text'
+ARCHIVE_GAME_DATA_FILES='cardid.tom data intro languageinis missions startopia.jpg text'
 
 CONFIG_FILES='./startopia.ini'
 DATA_DIRS='./profiles'
@@ -68,7 +63,6 @@ DATA_FILES='./*.txt'
 APP_MAIN_TYPE='wine'
 APP_MAIN_EXE='startopia.exe'
 APP_MAIN_ICON='startopia.exe'
-APP_MAIN_ICON_RES='32'
 
 PACKAGES_LIST='PKG_DATA PKG_BIN'
 
@@ -77,50 +71,52 @@ PKG_DATA_DESCRIPTION='data'
 
 PKG_BIN_ID="$GAME_ID"
 PKG_BIN_ARCH='32'
-PKG_BIN_DEPS_DEB="$PKG_DATA_ID, wine32-development | wine32 | wine-bin | wine-i386 | wine-staging-i386, wine:amd64 | wine"
-PKG_BIN_DEPS_ARCH="$PKG_DATA_ID wine"
+PKG_BIN_DEPS="$PKG_DATA_ID wine"
 
 # Load common functions
 
-target_version='2.0'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
-
-PKG='PKG_BIN'
-organize_data 'GAME_BIN'   "$PATH_GAME"
-
-PKG='PKG_DATA'
-organize_data 'DOC1'      "$PATH_DOC"
-organize_data 'DOC2'      "$PATH_DOC"
-organize_data 'GAME_DATA' "$PATH_GAME"
-
-PKG='PKG_BIN'
-extract_and_sort_icons_from 'APP_MAIN'
-move_icons_to 'PKG_DATA'
-
+prepare_package_layout
 rm --recursive "$PLAYIT_WORKDIR/gamedata"
+
+# Get game icon
+
+PKG='PKG_BIN'
+icons_get_from_package 'APP_MAIN'
+icons_move_to 'PKG_DATA'
 
 # Write launchers
 
 PKG='PKG_BIN'
-write_launcher 'APP_MAIN'
+launchers_write 'APP_MAIN'
 
 # Build package
